Address some PR comments + clean up ENS reducer
