+++ common/reducers/ens/index.js
@@ -1,0 +1,7 @@
+// @flow
+import domainSelector from './domainSelector';
+import domainRequests from './domainRequests';
+
+import { combineReducers } from 'redux';
+
+export const reducer = combineReducers({ domainSelector, domainRequests });
