+++ common/reducers/ens/types.js
@@ -1,0 +1,5 @@
+type Domain = {
+  state: string,
+  data: Object
+};
+export type State = { [string]: Domain };
